Rename file name and clean up code
